redis

+++ srcs/requirements/wordpress/tools/entrypoint.sh
@@ -1,7 +1,16 @@
 #!/bin/sh
 
-wp config create --dbname=$WORDPRESS_DB_NAME --dbuser=$WORDPRESS_DB_USER --dbpass=$WORDPRESS_DB_PASSWORD --dbhost=$WORDPRESS_DB_HOST --dbcharset="utf8" --dbcollate="utf8_general_ci" --allow-root
-wp core install --url=localhost --title=yo --admin_user=root --admin_password=$WORDPRESS_DB_PASSWORD --admin_email=vfdsafsfahfkl@student.s19.be --skip-email --allow-root
-wp user create vrogiste vrogiste@student.s19.be --role=author --user_pass=$WORDPRESS_DB_PASSWORD --allow-root
+if [ ! -f wp-config.php ]; then
+    wp config create --dbname=$WORDPRESS_DB_NAME --dbuser=$WORDPRESS_DB_USER --dbpass=$WORDPRESS_DB_PASSWORD --dbhost=$WORDPRESS_DB_HOST --dbcharset="utf8" --dbcollate="utf8_general_ci" --allow-root
+    wp core install --url=localhost --title=yo --admin_user=root --admin_password=$WORDPRESS_DB_PASSWORD --admin_email=vfdsafsfahfkl@student.s19.be --skip-email --allow-root
+    wp user create vrogiste vrogiste@student.s19.be --role=author --user_pass=$WORDPRESS_DB_PASSWORD --allow-root
 
+    sed -i wp-config.php "40i define( 'WP_REDIS_HOST', 'redis' );"
+    sed -i wp-config.php "41i define( 'WP_REDIS_PORT', 6379 );"
+    sed -i wp-config.php "42i define( 'WP_REDIS_TIMEOUT', 1 );"
+    sed -i wp-config.php "43i define( 'WP_REDIS_READ_TIMEOUT', 1 );"
+    sed -i wp-config.php "44i define( 'WP_REDIS_DATABASE', 0 );\n"
+    wp plugin install redis-cache --activate --allow-root
+    wp plugin update --all --allow-root
+fi
 exec $@
